This is for driving forward 1 second, back for two seconds, and forward two seconds.

+++ Stronghold2016/src/org/usfirst/frc/team340/robot/commands/DriveForward.java
@@ -11,7 +11,7 @@ public class DriveForward extends Command {
 	
 	private double speed;
 	
-    public DriveForward(int time, double speed) {
+    public DriveForward(double time, double speed) {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	requires(Robot.drive);
